requirements.txt
ircrobots ~= 0.6.1
pyyaml    ~= 6.0
aiohttp   == 3.8.1